+++ module/herosystem6e.mjs
@@ -32,6 +32,7 @@ import { CreateHeroCompendiums } from "./heroCompendiums.mjs";
 import "./utility/chat-dice.mjs";
 
 import "./testing/testing-main.mjs";
+import { EffectsPanel } from "./effects-panel.mjs";
 
 Hooks.once("init", async function () {
     game.CreateHeroCompendiums = CreateHeroCompendiums;
@@ -83,10 +84,6 @@ Hooks.once("init", async function () {
     CONFIG.statusEffects = HeroSystem6eActorActiveEffects.getEffects(HEROSYS.module);
 
     CONFIG.ActiveEffect.documentClass = HeroSystem6eActorActiveEffects;
-    CONFIG.ui.combat = HeroSystem6eCombatTracker;
-
-    CONFIG.ui.items = HeroSystem6eItemDirectory;
-    CONFIG.ui.compendium = HeroSystem6eCompendiumDirectory;
 
     CONFIG.Canvas.rulerClass = HeroRuler;
 
@@ -145,9 +142,38 @@ Hooks.once("init", async function () {
         `systems/${HEROSYS.module}/templates/item/item-partial-common.hbs`,
         `systems/${HEROSYS.module}/templates/actor/actor-sheet.hbs`,
         `systems/${HEROSYS.module}/templates/sidebar/partials/document-partial.hbs`,
+        `systems/${HEROSYS.module}/templates/system/effects-panel.hbs`,
     ];
     // Handlebars Templates and Partials
     loadTemplates(templatePaths);
+});
+
+Hooks.once("init", () => {
+    // Assign the Sidebar subclasses
+    //CONFIG.ui.actors =
+    CONFIG.ui.items = HeroSystem6eItemDirectory;
+    CONFIG.ui.combat = HeroSystem6eCombatTracker;
+    //CONFIG.ui.chat =
+    CONFIG.ui.compendium = HeroSystem6eCompendiumDirectory;
+    //CONFIG.ui.hotbar =
+
+    // Insert templates into DOM tree so Applications can render into
+    if (document.querySelector("#ui-top") !== null) {
+        // Template element for effects-panel
+        const uiTop = document.querySelector("#ui-top");
+        const template = document.createElement("template");
+        template.setAttribute("id", "hero-effects-panel");
+        uiTop?.insertAdjacentElement("afterend", template);
+    }
+    game[HEROSYS.module] ??= {};
+    game[HEROSYS.module].effectPanel = new EffectsPanel();
+});
+
+Hooks.on("canvasReady", () => {
+    // Effect Panel singleton application
+    game[HEROSYS.module].effectPanel.render(true);
+    if (!canvas.scene) return;
+    if (game.ready) canvas.scene.reset();
 });
 
 Hooks.once("ready", async function () {